+++ tests/plugins/variants/startingStyleVariant.test.js
@@ -1,0 +1,9 @@
+import { css, quickVariantPluginTest } from '../../util/run'
+
+quickVariantPluginTest('startingStyleVariant').toMatchFormattedCss(css`
+  @starting-style {
+    .starting\:flex {
+      display: flex;
+    }
+  }
+`)